use renamed plugin protocol dependency
