+++ go.mod
@@ -12,7 +12,7 @@ require (
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.25.0
 	github.com/batchcorp/pgoutput v0.3.2
 	github.com/bufbuild/buf v1.0.0
-	github.com/conduitio/connector-plugin v0.0.0-20220215124418-565125202e8d
+	github.com/conduitio/conduit-plugin-protocol v0.0.0-20220301130721-cbefccd064ea
 	github.com/dgraph-io/badger/v3 v3.2103.2
 	github.com/dop251/goja v0.0.0-20210225094849-f3cfc97811c0
 	github.com/golang/mock v1.6.0
@@ -35,7 +35,7 @@ require (
 	github.com/segmentio/kafka-go v0.4.28
 	github.com/xitongsys/parquet-go v1.6.0
 	github.com/xitongsys/parquet-go-source v0.0.0-20201108113611-f372b7d813be
-	go.buf.build/library/go-grpc/conduitio/connector-plugin v1.4.1
+	go.buf.build/library/go-grpc/conduitio/conduit-plugin-protocol v1.4.1
 	golang.org/x/tools v0.1.9
 	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1
 	google.golang.org/genproto v0.0.0-20220211171837-173942840c17
