+++ pkg/plugin/builtin/v1/internal/toplugin/record.go
@@ -15,9 +15,9 @@
 package toplugin
 
 import (
+	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/record"
-	"github.com/conduitio/connector-plugin/cpluginv1"
 )
 
 func Record(in record.Record) (cpluginv1.Record, error) {
