+++ pkg/plugin/sdk/destination.go
@@ -21,8 +21,8 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
-	"github.com/conduitio/connector-plugin/cpluginv1"
 )
 
 // Destination receives records from Conduit and writes them to 3rd party
