+++ pkg/plugin/sdk/source.go
@@ -20,8 +20,8 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
-	"github.com/conduitio/connector-plugin/cpluginv1"
 	"github.com/jpillora/backoff"
 	"gopkg.in/tomb.v2"
 )
