+++ pkg/plugin/sdk/specifier.go
@@ -17,7 +17,7 @@ package sdk
 import (
 	"context"
 
-	"github.com/conduitio/connector-plugin/cpluginv1"
+	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
 )
 
 // Specification is returned by a plugin when Specify is called.
