+++ pkg/plugin/standalone/v1/internal/fromproto/source.go
@@ -16,7 +16,7 @@ package fromproto
 
 import (
 	"github.com/conduitio/conduit/pkg/record"
-	connectorv1 "go.buf.build/library/go-grpc/conduitio/connector-plugin/connector/v1"
+	connectorv1 "go.buf.build/library/go-grpc/conduitio/conduit-plugin-protocol/connector/v1"
 )
 
 func SourceRunResponse(in *connectorv1.Source_Run_Response) (record.Record, error) {
