+++ pkg/plugin/standalone/v1/internal/fromproto/specifier.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 
 	"github.com/conduitio/conduit/pkg/plugin"
-	connectorv1 "go.buf.build/library/go-grpc/conduitio/connector-plugin/connector/v1"
+	connectorv1 "go.buf.build/library/go-grpc/conduitio/conduit-plugin-protocol/connector/v1"
 )
 
 func SpecifierSpecifyResponse(in *connectorv1.Specifier_Specify_Response) (plugin.Specification, error) {
