+++ pkg/plugin/standalone/v1/internal/toproto/destination.go
@@ -16,7 +16,7 @@ package toproto
 
 import (
 	"github.com/conduitio/conduit/pkg/record"
-	connectorv1 "go.buf.build/library/go-grpc/conduitio/connector-plugin/connector/v1"
+	connectorv1 "go.buf.build/library/go-grpc/conduitio/conduit-plugin-protocol/connector/v1"
 )
 
 func DestinationConfigureRequest(in map[string]string) (*connectorv1.Destination_Configure_Request, error) {
